utils/labmanager/requirements.txt
jq
pydo
configparser
prettytable
requests
pyyaml